+++ engine/common/components/src/main/scala/pl/touk/nussknacker/engine/common/components/DecisionTable.scala
@@ -1,17 +1,15 @@
 package pl.touk.nussknacker.engine.common.components
 
+import cats.data.NonEmptyList
 import pl.touk.nussknacker.engine.api._
-import pl.touk.nussknacker.engine.api.context.ValidationContext
-import pl.touk.nussknacker.engine.api.context.transformation.{
-  DefinedEagerParameter,
-  NodeDependencyValue,
-  SingleInputDynamicComponent
-}
+import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
+import pl.touk.nussknacker.engine.api.context.transformation._
 import pl.touk.nussknacker.engine.api.definition._
+import pl.touk.nussknacker.engine.api.generics.ExpressionParseError.TabularDataDefinitionParserErrorDetails
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.process.ComponentUseCase
 import pl.touk.nussknacker.engine.api.test.InvocationCollectors
-import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedObjectTypingResult, TypingResult}
+import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedObjectTypingResult, TypingResult, Unknown}
 import pl.touk.nussknacker.engine.graph.expression.TabularTypedData
 import pl.touk.nussknacker.engine.graph.expression.TabularTypedData.Column
 
@@ -38,15 +36,16 @@ object DecisionTable extends EagerService with SingleInputDynamicComponent[Servi
   private object MatchConditionTableExpressionParameter {
     val name: ParameterName = ParameterName("Match condition")
 
-    val declaration: ParameterCreator[TabularTypedData] with ParameterExtractor[LazyParameter[lang.Boolean]] = {
+    val declaration
+        : ParameterCreator[Iterable[Column.Definition]] with ParameterExtractor[LazyParameter[lang.Boolean]] = {
       ParameterDeclaration
         .lazyMandatory[java.lang.Boolean](name)
-        .withAdvancedCreator[TabularTypedData](
-          create = data =>
+        .withAdvancedCreator[Iterable[Column.Definition]](
+          create = columnDefinitions =>
             _.copy(additionalVariables =
               Map(
                 decisionTableRowRuntimeVariableName -> AdditionalVariableProvidedInRuntime(
-                  rowDataTypingResult(data.columnDefinitions)
+                  rowDataTypingResult(columnDefinitions)
                 )
               )
             )
@@ -62,7 +61,7 @@ object DecisionTable extends EagerService with SingleInputDynamicComponent[Servi
   override def contextTransformation(context: ValidationContext, dependencies: List[NodeDependencyValue])(
       implicit nodeId: NodeId
   ): ContextTransformationDefinition = {
-    prepare orElse basicDecisionTableParameterReady orElse allParametersReady(context, dependencies)
+    prepare orElse basicDecisionTableParameterDefined orElse allParametersDefined(context, dependencies)
   }
 
   override def implementation(
@@ -83,13 +82,54 @@ object DecisionTable extends EagerService with SingleInputDynamicComponent[Servi
     )
   }
 
-  private lazy val basicDecisionTableParameterReady: ContextTransformationDefinition = {
+  private lazy val basicDecisionTableParameterDefined: ContextTransformationDefinition = {
     case TransformationStep((name, DefinedEagerParameter(data: TabularTypedData, _)) :: Nil, _)
         if name == BasicDecisionTableParameter.name =>
       NextParameters(
-        parameters = MatchConditionTableExpressionParameter.declaration.createParameter(data) :: Nil,
+        parameters = MatchConditionTableExpressionParameter.declaration.createParameter(data.columnDefinitions) :: Nil,
         errors = List.empty,
         state = None
+      )
+    case TransformationStep((name, FailedToDefineParameter(errors)) :: Nil, _)
+        if name == BasicDecisionTableParameter.name =>
+      val columnDefinitions = extractColumnDefinitionsFrom(errors).getOrElse(TabularTypedData.empty.columnDefinitions)
+      NextParameters(
+        parameters = MatchConditionTableExpressionParameter.declaration.createParameter(columnDefinitions) :: Nil,
+        errors = List.empty,
+        state = None
+      )
+  }
+
+  private def allParametersDefined(context: ValidationContext, dependencies: List[NodeDependencyValue])(
+      implicit nodeId: NodeId
+  ): ContextTransformationDefinition = {
+    case TransformationStep(
+          (firstParamName, FailedToDefineParameter(_)) ::
+          (secondParamName, _) :: Nil,
+          _
+        )
+        if firstParamName == BasicDecisionTableParameter.name &&
+          secondParamName == MatchConditionTableExpressionParameter.name =>
+      FinalResults.forValidation(context)(
+        _.withVariable(
+          name = OutputVariableNameDependency.extract(dependencies),
+          value = Unknown,
+          paramName = None
+        )
+      )
+    case TransformationStep(
+          (firstParamName, DefinedEagerParameter(data: TabularTypedData, _)) ::
+          (secondParamName, _) :: Nil,
+          _
+        )
+        if firstParamName == BasicDecisionTableParameter.name &&
+          secondParamName == MatchConditionTableExpressionParameter.name =>
+      FinalResults.forValidation(context)(
+        _.withVariable(
+          name = OutputVariableNameDependency.extract(dependencies),
+          value = componentResultTypingResult(data.columnDefinitions),
+          paramName = None
+        )
       )
   }
 
@@ -126,25 +166,6 @@ object DecisionTable extends EagerService with SingleInputDynamicComponent[Servi
 
   }
 
-  private def allParametersReady(context: ValidationContext, dependencies: List[NodeDependencyValue])(
-      implicit nodeId: NodeId
-  ): ContextTransformationDefinition = {
-    case TransformationStep(
-          (firstParamName, DefinedEagerParameter(data: TabularTypedData, _)) ::
-          (secondParamName, _) :: Nil,
-          _
-        )
-        if firstParamName == BasicDecisionTableParameter.name &&
-          secondParamName == MatchConditionTableExpressionParameter.name =>
-      FinalResults.forValidation(context)(
-        _.withVariable(
-          name = OutputVariableNameDependency.extract(dependencies),
-          value = componentResultTypingResult(data.columnDefinitions),
-          paramName = None
-        )
-      )
-  }
-
   private def componentResultTypingResult(columnDefinitions: Iterable[Column.Definition]): TypingResult = {
     Typed.genericTypeClass(classOf[Output], rowDataTypingResult(columnDefinitions) :: Nil)
   }
@@ -155,5 +176,20 @@ object DecisionTable extends EagerService with SingleInputDynamicComponent[Servi
         columnDef.name -> Typed.typedClass(columnDef.aType)
       }.toMap
     )
+
+  private def extractColumnDefinitionsFrom(errors: NonEmptyList[ProcessCompilationError]) = {
+    errors.toList
+      .flatMap {
+        case e: ProcessCompilationError.ExpressionParserCompilationError =>
+          e.details match {
+            case Some(details: TabularDataDefinitionParserErrorDetails) => Some(details)
+            case Some(_) | None                                         => None
+          }
+        case _ =>
+          None
+      }
+      .map(_.columnDefinitions.map(cd => Column.Definition(cd.name, cd.aType)))
+      .headOption
+  }
 
 }